Backend/flask_app/app/requirements.txt
alembic==1.4.1
amqp==2.5.2
bidict==0.21.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
dnspython==1.16.0
dominate==2.6.0
eventlet==0.25.1
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
greenlet==0.4.15
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
kombu==4.6.8
Mako==1.1.2
MarkupSafe==1.1.1
mock==4.0.2
monotonic==1.5
path==13.1.0
path.py==12.4.0
polling==0.3.1
PyJWT==1.4.2
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==4.0.0
python-socketio==5.0.4
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
urllib3==1.25.8
vine==1.3.0
visitor==0.1.3
Werkzeug==1.0.0
wrapt==1.12.1
WTForms==2.3.3
zipp==3.1.0

PyQRCode~=1.2.1
onetimepass~=1.0.1
fpdf~=1.7.2
PyPDF2~=1.26.0
rq~=1.9.0
redis~=3.5.3
PyLaTeX~=1.4.1
